requirements.txt
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
jinja2>=3.1.0
feedparser>=6.0.0
